Move compat tests under webgpu
Now that the compat tests skip in non-compat
it seems easiest if they're just included under webgpu:
as then other testing infrastructure does not have
to treat them differently.
