+++ src/webgpu/compat/api/validation/render_pipeline/shader_module.spec.ts
@@ -2,7 +2,7 @@ export const description = `
 Tests limitations of createRenderPipeline related to shader modules in compat mode.
 `;
 
-import { makeTestGroup } from '../../../../common/framework/test_group.js';
+import { makeTestGroup } from '../../../../../common/framework/test_group.js';
 import { CompatibilityTest } from '../../../compatibility_test.js';
 
 export const g = makeTestGroup(CompatibilityTest);
